server/requirements.txt
fastapi
uvicorn[standard]
opencensus-ext-azure
python-dotenv
httpx
pydantic
google-api-python-client
google-auth-httplib2
google-auth